% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{PhysioEMG-package}
\alias{PhysioEMG}
\alias{PhysioEMG-package}
\title{PhysioEMG: EMG Analysis Functions for PhysioExperiment Objects}
\description{
Provides electromyography (EMG) analysis functions for PhysioExperiment objects. Includes envelope extraction (RMS, Hilbert), muscle activation onset detection (Hodges-Bui, Teager-Kaiser), fatigue analysis (median frequency shift), and muscle synergy decomposition (NMF, PCA, ICA).
}
\examples{
# Simulate EMG, extract an RMS envelope, and decompose into synergies
pe <- make_emg(n_time = 2000, n_channels = 4, sr = 1000)
pe <- emgEnvelope(pe, method = "rms")
muscleSynergy(pe, n_synergies = 2, method = "nmf", seed = 1,
              assay_name = "envelope")$vaf
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/x-biosignal/PhysioEMG}
  \item \url{https://x-biosignal.r-universe.dev/PhysioEMG}
  \item \url{https://x-biosignal.github.io/PhysioEMG/}
  \item Report bugs at \url{https://github.com/x-biosignal/PhysioEMG/issues}
}

}
\author{
\strong{Maintainer}: Yusuke Matsui \email{mail.to.matsui@gmail.com}

}
\keyword{internal}
