% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emg-plot.R
\name{plotEmgOnset}
\alias{plotEmgOnset}
\title{Plot EMG Activation Onsets and Offsets}
\usage{
plotEmgOnset(x, channel = 1, method = "hodges_bui", assay_name = NULL, ...)
}
\arguments{
\item{x}{A PhysioExperiment object with EMG data.}

\item{channel}{Channel index or label to plot (default: 1).}

\item{method}{Onset method passed to \code{\link[=emgOnsetDetect]{emgOnsetDetect()}} (default:
"hodges_bui").}

\item{assay_name}{Input assay name (default: first assay).}

\item{...}{Further arguments for \code{\link[=emgOnsetDetect]{emgOnsetDetect()}}.}
}
\value{
A ggplot object.
}
\description{
Plots the rectified signal with detected activation intervals shaded and
onset/offset times marked.
}
\examples{
pe <- make_emg_contraction()
plotEmgOnset(pe)
}
\references{
Hodges, P.W. & Bui, B.H. (1996). "A comparison of computer-based
methods for the determination of onset of muscle contraction using
electromyography." Electroencephalography and Clinical Neurophysiology,
101(6), 511-519. doi:10.1016/S0921-884X(96)95190-5
}
\seealso{
\code{\link[=emgOnsetDetect]{emgOnsetDetect()}}
}
