% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{PhysioEMG-package}
\alias{PhysioEMG}
\alias{PhysioEMG-package}
\title{PhysioEMG: EMG Analysis Functions for PhysioExperiment Objects}
\description{
Provides electromyography (EMG) analysis functions for PhysioExperiment objects. Includes envelope extraction (RMS, Hilbert), muscle activation onset detection (Hodges-Bui, Teager-Kaiser), fatigue analysis (median frequency shift), and muscle synergy decomposition (NMF, PCA, ICA).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/x-biosignal/PhysioEMG}
  \item \url{https://x-biosignal.r-universe.dev/PhysioEMG}
  \item \url{https://x-biosignal.github.io/PhysioEMG/}
  \item Report bugs at \url{https://github.com/x-biosignal/PhysioEMG/issues}
}

}
\author{
\strong{Maintainer}: Yusuke Matsui \email{mail.to.matsui@gmail.com}

}
\keyword{internal}
