% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emg-decompose.R
\name{synergyReconstruct}
\alias{synergyReconstruct}
\title{Reconstruct Data from Synergies}
\usage{
synergyReconstruct(synergy_result, n_synergies)
}
\arguments{
\item{synergy_result}{Result from \code{\link{muscleSynergy}}.}

\item{n_synergies}{Number of synergies to use for reconstruction.}
}
\value{
A list with:
\itemize{
\item \code{reconstructed}: Reconstructed data matrix
\item \code{vaf}: VAF of the reconstruction
}
}
\description{
Reconstructs EMG data using a subset of synergies.
}
\references{
De Luca, C.J. (1997). "The use of surface electromyography in
biomechanics." Journal of Applied Biomechanics, 13(2), 135-163.
doi:10.1123/jab.13.2.135
}
\seealso{
\code{\link[=muscleSynergy]{muscleSynergy()}} for computing the initial decomposition,
\code{\link[=synergyCompare]{synergyCompare()}} for comparing synergy solutions
}
